Package/jobassist_unpacked/usr/share/jobassist/requirements.txt
psycopg2-binary
Flask
gunicorn
flask-sqlalchemy
flask-cors
flask
python-magic
